Start of actual code
Using tkinter for now, since that gives me basic UI in a pretty simple
manner. Took some finagling to get about what I want, but I have the
basics in place now to type in stuff, hit enter, and respond. Next up is
to see about threading so that I can actually have the timer count down
and so forth, but we'll see how that goes. Hopefully that's a kosher
thing to do to a Tk application; if not, we'll need to figure out how to
do tk timers instead (bleah). Or maybe Tk named events. Or something.