Basic gameplay now there
Fleshed out the basic idea of how the system is supposed to work. I'm
not sure if it's fast enough or if it's too fast, but the basics are
there. I'm going to get Val's opinion on the matter next and see if we
need to course correct anything from that point.